Rename "train / real_train" to conventional "train / test"
As can be seen, in the old training_scripts/prepreocess_knees.py,
the code stored an intermediate copy of the test set as "real_train".
This intermediate copy is loaded in the various evaulation scripts.
This isn't technically wrong and doesn't affect the results, but could be a source
of some distress.